requirements.txt
streamlit
spacy
torch
transformers
python-docx
pyPDF2
faiss-cpu
nltk
langchain
sentence-transformers
langchain-community


